Fix Dockerfile file permissions and UID handling
Removed the user creation from the Dockerfile. Changed the
file permissions to be only for a group.

The design philosophy is that "You can run the container
with root user. There is nothing that prevents this."

The most important parts container security hardening must be
done in the host that runs the container images. For example
the UID should come from the host in order to manage volume
mounts and make sure that the UID that runs the container is
not a root user. It is useless try to to circumvent this inside
the container.

The container user is always part of the root group. No matter
if there is no UID given from host, UID defined from host or
how ever you want to start the container.  Because of this, the
file privileges are always set for a group.

USER is not currently set in the container. This may be a
problem with OpenShift [1]. But based on the above, it is not
clear for the author why OpenShift requires this.

[1] https://github.com/openshift/openshift-docs/blob/master/creating_images/guidelines.adoc

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>