Fix running tests with Tox
Tox now runs tests against correct Python version. The
problem likely was syntax 'pip install ..' when it hast
to be 'python -m pip ..' and 'python -m pytest ..' to
install the dependencies and run tests against correct
Python version in virtual environment that Tox installs.

Use Makefile and make when running tests. All tools must
use the make in order to guarantee same parameters when
tests are run. This also helps with maintenance where
different tools (Travis CI, Tox) do not have different
commands that are not remembered to updated at once.

Add install-test target to reduce the dependencies that
are installed for test.

There is a problem with Python 3.4 that requires the
zlib. This requires compiling the Python 3.4 locally
which does not work yet. See The tox.ini documentation
for more information.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>