Fix Dockerfile USER and update the document
The USER must be set in the Dockerfile. The reason is that
some infrastructures like OpenShift use the USER instruction
in the Dockerfile to determine if the container is run with
a root user [1]. The more detaield reasons can be found
from the documentation part of the Dockerfile.

Fix the Dockerfile documentation after recent changes. Now
the Dockerfile should be configured so that user can do
security measurement from host if wanted without breaking
normal unsecure use cases.

The area is complicated so this is to be seen how it goes.

[1] https://docs.openshift.com/container-platform/3.11/creating_images/guidelines.html

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>