Fix allocating non-root GID for server container
Previously when the server was started with other than root
GID, the server startup failure. This is possible use case
that is supported on top of changing the UID.

This required adding file permissions for 'others' as well
as 'groups' for the server. See more details from the updated
Dockerfile documentation.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>