Add CHANGELOG entry for server in container fixes
The Snippy server Dockefile configuration was completely
refactored and most of the use cases are now tested. The
container should now run with the examples provided in the
Dockefile. The intention is to move the documentation in
the Dockerfile at some point to user documentation when
implementation stabilizes. As of now it will stay for a
while in the Dockerfile itself.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>