Add Python 3.7 for Travis CI
Also the Travis CI YAML configuration file was refactored.
There used to be a core dump which was the reason to have
debug settings in the Travis YAML. Core dumps have not been
visible and the related settings were removed.

Also a Makefile based install and run commands were used.
There is new 'test-vi' target that does not include server
and docker tests. Those tests are using external processes
and they do not contribute to code test coverage.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>