Edit Python module versions for release
Falcon was upgraded to 2.0.0. This also revealed a bug in
few tests. The test added trailing slash at the end of the
URI. With Falcon 2.0.0 this generates 404 Not Found error.
The reason is is because /path/ and /path have different
semantics which is import in URLs [1].

This seems to be because of 'RequestOptions.strip_url_path
_trailing_slash' [2].

[1] https://stackoverflow.com/a/16445016
[2] https://github.com/falconry/falcon/blob/master/CHANGES.rst

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>