Fix requirements.txt
The requirements.txt must contain only the dependencies that
are needed by the Snippy server installation. Dependencies
of dependencies are not listed in this file. This makes the
outdated version checks with the PyUp more accurate because
it does not complain outdated modules in third party modules
but in this project only.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>