Remove parallel execution of tests in Travis CI
Run tests with multiple cores in Travis CI actually slows
down the test tun time (test-ci). There are only two cores
in Travis CI build environments. It may be that the slow
setup for the multicore processing is slow with few cores.

With sequential the tests were run on average 24 seconds
on Travis CI xenial environments. With parallel this was
around 32-34 seconds.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>