Fix storage usage in case of healthcheck
The server healthcheck does not need the storage anymore.
Previously the storage and database was connected even
when a server healthcheck was made. This is a risk from
robustness and security point of view. It may be that
constant healtcheck which also constantly opens and
closes the database, may cause leaks or performance
issues.

The main design approach is still that the code flows
from start to end and there are no premature exit on
the road to end. Because this, the code location where
the healthcheck is run is still at the end of the code.
The database connection is just silently bypassed.

A successful healthcheck itself must not produce any
logs. This avoid flooding the log systems with constant
flow of logs. The healthcheck will use API endpoints
which can generate logs if logs are enabled.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>