Change server healthcheck API to /hello
The healtcheck is now sent to the server /hello API endpoint
instead to server base path. The /hello allows easy separation
of the healtcheck requests.

As of now, there must be no logs printed from periodic health
that is originated from the same host where the server is
running. The purpose is to avoid flooding the logs with the
healthcheck. If there is a problem that goes unnoticed for a
while, the constant healtcheck could roll the logs and the
original problem would be lost. As of now, there are no logs
at all when a successful healthcheck is run.

If the healtcheck came from external IP that is not the
server IP, the HTTP request in the /hello API endpoint
is logged normally if logs are activated. This is to trace
possible malicious users sending large amount of requests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>