Edit Dockerfile documentation on how UID is chosen
There was a poorly worded chapter that gave an impression
that a Dockerfile and the USER instruction can override
the UID value set from the host. This is not the case. When
a host and its dockerd set the UID (and GID) explicitly, it
will always override the value in Dockerfile. Only if host
did not explicitly set the UID, Dockerfile and container
can override the value.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>