Fix error message from duplicated content
If the content default was imported again when the server
restarted, there was an unnecessary error log. The log was
from the unique constraint violation when content with
already existing UUID (or digest) was added.

If the defaults are being imported with '--defaults' option,
there are no more info logs from the exception.

If the exceptions happens when the '--default' option is
not used, the same info level log with exception details
is printed than before.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>