Add Bash completion for Snippy
There is now Bash completion script for the command line
interface. Currently the activation is manual. The server
command (to be implemented) is not autocompleted. This
would complicate the 'search' command by forcing user to
write 'sea<tab>' to get the most commonly used command
completed.

The location where the script must be set can be checked
with command [1]:

  $ pkg-config --variable=completionsdir bash-completion

The script can be enabled with:

  $ sudo cp snippy.bash-completion /etc/bash_completion.d/
  $ exec bash

[1] https://github.com/scop/bash-completion

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>