Add 'server' operation for command line interface
There is now a 'server' operation that is a positional
argument with other operations like 'create' or 'search'.

The server operation is hidden on purpose. User just have
to know that it exists. The main scope of the tool from the
command line is the command line version.

Code determines from the mandatory option '--server-host'
if the server is run, not from the 'server'. This is mostly
due to previous implementation. There was no real reason
seen why the existing code would have to transformed to
'server' option since the mandatory '--server-host' is
always used. It could be possible to argue this reason
from code maintainability point of view since this is
party a surprise that the 'server' is there.

One reasons to have a 'server' is the Bash completion. This
suits very well for that and it makes the CLI more logical.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>