Change content category command line options
Removed options '--snippet', '--solution', '--reference' and
'--all'. These were redundant options. The content category
is now always defined with the '--scat' option.

The '--scat' option supports single category or multiple
categories with same syntax that any of options with list
context.

The removed options were confusing and redundant. They were
not very well suited for example for Bash completion and the
CLI should be easier to understand with only one content
category option.

Fixed configuration source initialized. The config source
base class accidentally called the init_conf() when the
same method is called by Cli() and Api(). This initialized
the configuration twice. There may have been 5-10% boost
in performance and 5->4ms latency improvement due to this.
The problem is that the references were run on Fedora 26
when the new devel machine is Fedora 30 which is a bit
smoother.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>