Fix operation filename '--file' option
There was an internal failure where the '--file' option
was stored to content 'filename' attribute. These are
different attributes. The failure was likely not visible
externally.

Now there are 'operation_file' and 'filename' attributes
in configuration source for these two attributes. The
'operation_file' contains the value from '--file' option
and the 'filename' is the resource attribute 'filename'.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>