Fix development environment installation document
There are now better instructions how to set a CPython and
PyPy Python test environments. The Makefile was fixed to
support '--user' install as well as to work with CPython
and PyPy in a virtual environment with same commands.

The virtual environments work nicely with CPython and
PyPy virtual environments because the 'python' always
points to correct binary inside a virtual environment.

The '-m ensurepip' is used and needed in the 'upgrade-
wheel' target because the PyPy does not install pip by
default. The 'ensurepip' is needed to install the pip.
In case of CPython, the command just passes because the
pip is already installed.

Changed also the default 'make test' to be the one that
runs tests on parallel and does not execute server and
docker tests. The 'make test' is the default test to be
run for example with CI. The parallel execution can be
disabled with Python 3 with:

  $ make test PYTEST_CORES=

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>