Add PyPy2 and PyPy3 for Travis CI
The tests are now executed with PyPy as well. This means
the 'test' target in make that does not run server or
Docker tests. The server test is not working for reason
with PyPy and the Docker tests require Docker in Docker
setup that may not be supported by Travis CI.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>