Fix server test cases with PyPy
The server tests failed because the server startup was a bit
slower with PyPy than with CPython. There was a 1 second wait
that was not enough with PyPy to wait the server startup.

Now there is a proper while based wait that polls the hello
API endpoint to see when the server is up and running. This
speeds up the testing a bit since usually the server is up
and running in less than second.

Also the server process creation was moved to Pytest fixture
that allows releasing the server in case of failure. Previously
the server was left running in case of failure.

The port allocation for the server is now completely inside
the conftest.py module. This allows dynamic port allocation
to allow running multiple servers in parallel. The next step
is to make also the server tests to support parallel tests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>