Add in-memory database with Sqlite
It is now possible to use 'in-memory' with --storage-type.
This option makes the Sqlite database to operate only in
memory. All updates are lost when the process restarts.

The use case is parallel execution of server tests. This
requires running multiple servers in parallel and the
in-memory database is a nice way to achieve this.

This may be useful also in scenarios where only default
content is required. Because of this, this is a publicly
available use case.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>