Add parallel server tests
The tests with a real REST API server can be now run in
parallel. The port is allocated by operating system and
the tests just use a handle to the HTTP session where to
send messages. The Pytest fixtures will take case of
creating and releasing the server process.

The server startup is based on polling specific log and
hello API endpoint. The specific server startup log is
needed to find out the randomly allocated ephemeral port.
The reason to use logs is to avoid using 'psutil' Python
module and operating system specific commands to find out
the created process and port for example based on PID.
Also the psutil module requires C compiler which is
always cumbersome and error prone with Python.

Because the logs are used for the server starting, it
means that using the quiet option '-q' does not work
with the server tests.

Also because the logs are read during the server startup,
tests cases using the 'process' fixture are not able to
read the server startup logs before and including the
'snippy server running at ...' log. The logs are read
from the stdout and they get removed when read.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>