Fix project test dependency installation
The Pytest requires pluggy. But for some reason that is not
know, the pytest does not install correct version from pluggy.
The pluggy was previously installed separately for example for
Travis CI. Now the pluggy is explicitly installed in the test
requirements for the project.

The mock and Pytest packages were updated.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>