Fix command line usage with Docker container
There was a fault that prevented using the container like a
command line tool. The ENTRYPOINT in Dockerfile contained the
'server' argument which invalidated for example the 'search'
argument given when container was started.

Now both server and command line usage are working. This was
deprecated because Docker tests were not run in Fedora 30
previously.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>