Add PostgreSQL testing to Travis CI
The default testing password for PostgreSQL was changed to
empty string to match the Travis CI PostgreSQL default
password.

There is now the used database type in the Pytest report
header. The color tries to highlight it during tests and
it matches to the green color used in the Pytest.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>