Fix PostgreSQL install with Python 2.7 and PyPy2
The environment marker 'implementation_name' did not work
with Python 2.7 and PyPy2. The marker 'platform_python_
implementation' works with all Python versions when tried
with the latest setuptools. Both of these markers are from
PEP 508 [1].

Based on setuptools change history, the	'platform_python_
implementation' marker has been working since setuptools
18.5 (2015.11.01) [2]. The PEP 508 is mentioned in version
20.2 (2016.02.19) [2].

It is likely that Python 2 based implementations are so old
that only the older environment marker is working.

It is not know which is preferred and if there are some
real differences between these two markers.

[1] https://www.python.org/dev/peps/pep-0508/
[2] https://setuptools.readthedocs.io/en/latest/history.html

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>