Fix test tool configuration
Test tools Pylint, Flake8 and Bandit configurations were
simplified.

The Pylint configuration was a long default configuration
that contained all the options with most of the options
in default values. There was also a separate files for the
tests and snippy source code. Now only the needed options
are configured and the files are merged to one default
file 'pylintrc'.

The Flake8 configuration was moved to setup.cfg to remove
own configuration file.

The result files for Pylint and Bandit were removed. This
was not a good idea in the past.

The configuration should be read from pyproject.toml. But
the tools do not yet support this feature.

[1] https://github.com/PyCQA/pylint/issues/617
[2] https://gitlab.com/pycqa/flake8/issues/428
[3] https://github.com/pytest-dev/pytest/issues/1556

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>