Fix server tests and performance test results
There was a failure in server tests after previous commit
that changed the server startup log.

The server performance results were also updated. Server
process is now ran with in-memory database. This speed
up the reference performance tests so much that all test
results in the past are now invalid.

The reason to use the in-memory database is to allow
parallel testing of server tests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>