Fix Makefile upgrade-wheel with Fedora 26
The upgrade-wheel target uses 'python -m ensurepip' to make
sure that the pip is installed. This fails with Fedora 26
with Python 2.7 and Python 3.6 when pip is installed. This
works with Fedora 30. This likely fails with Fedora release
before Fedora 30.

Now the 'ensurepip' is ran only if the pip is not installed.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>