Fix CLI option tests for logging
The were exceptions from Python logging module. These were
caused when the Snippy.release() was called from the 'snippy'
fixture when the release() was already called. The CLI tests
run the Snippy from start to end and the release() is already
done when the Pytest fixture finalizer is done.

When the 'snippy' fixture is used, the snippy.release()
should not be called.

There is also the 'logging.raiseExceptions = False' in the
Logger() these days. This setting prevents exceptions from
the Python logging module.

The above clarified the long standing problems in the test
and allowed to enable tests that there are no output in
the stderr.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>