Fix failing tests if content imported
Two CLI option tests were failing if content was imported.

The test pattern that creates Snippy object in test suffers
from a problem that the Snippy object uses the real database
by default. If there are content imported into the Sqlite
database, it affects the tests.

Changed the failing tests to use fixtures that mock the
Sqlite database properly.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>