Fix updating content between formats
There is a problem where Markdown native solution cannot
be updated in text format. Now this is fixed so that the
Markdown formatted content can be updated in text format.

This does not change that actual data in Markdown format.
This just merely adds a text header and metadata. It would
be too complex to start to format the actual generic data
between text and Markdown format.

There is now metadata section also in the text formatted
solutions. The text solution header is now own header and
it is not part of the data.

Fixed parsing of solution description from the solution
template. There was a mix between 'Description' and
'description' which prevented parsing of the description
attribute.

Fixed --format option in export. Now the export obeys the
--format option.

This is work in progress and some tests are still failing.

[skip ci]

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>