Fix updating content between formats (breaking)
The Snippet and Reference content text templates now have
metadata. The reason is that the format change update
required to have all the content attributes in the text
template in order to manage the format transfer. Now the
same change is in Snippet and Reference templates.

The templates must contain attributes only once in order to
keep the implementation simple. Attributes presented many
times may create conflict what attributes to read. This also
increases testing effort.

This can be considered as a breaking change because the text
templates were changed. How ever, the functionality is likely
to be backward compatible. There are just no guarantees from
this.

Updated tests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>