Fix parsing metadata from solution text template
The metadata for solution was not parsed correct from the
text template metadata. The 'source', 'name' and 'versions'
attributes were not read from the template.

Changed solution text template metadata to have only each
attribute once in the template. This tries to simplify the
implementation and testing in order to avoid having one
attribute multiple times in the templates.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>