Fix REST API server response if HTTP request fails
The REST API server must send an error response if the
HTTP request is erroneous. The problem is that some of
the attributes are checked in JSON schema validation
when the HTTP request is received, and some of values
of attributes when a resource from request is processed.

Previously if the versions attribute contained incorrect
values, the HTTP request was accepted and the resource
was stored. This is now rejected with an error.

There are still error(s) in case multiple resources are
created with one request. If the first is successful but
the later ones will fail, the response is 201 and the
successful is stored. The problem is that resources are
processed one by one and the error check does not see
the following resources in collection when the first
resource is stored.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>