Fix validation of multiple resources in one request
If HTTP request was received with multiple resources to be
created, the error response was incorrect. If the first
resource was successfully created and any of the following
resources in the HTTP request missed mandatory attribute
'data' or 'links, the response was 201 Created when it
should have been an HTTP error.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>