Fix installation instructions for development
The development environment examples were updated. There
were errors that were corrected. For example it is not
advised to upgrade the system pip package. The upgrade
should be made with 'python -m pip ... --user' that is
updating the user specific package. Upgrading system
package may risk the stability of the OS.

The scripts to install, deploy and delete the virtual
environments were enhanced. Mainly there is now nicer
print from each step.

The Makefile targets were silenced by default. The
assumption is that the install and deploy scripts are
printing required status information about the install
process.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>