Fix development installation instructions
Python install process is complicated for the author and
there are no claims that this is perfect.

With current understanding: The development installation
uses python3. The used virtual environment modules are
installed with python3 under python module script dir.
This avoids changing the global settings if user has an
existing setup. This allows for example creating new user
for Snippy development that has own virtual module setup.

The instructions avoid adding anything into Python module
global. The virtual environment modules are in user module
script directory and pip, setuptools and wheel are installed
as part of each virtual environment.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>