Add sort option to CLI to sort based on attributes
It is now possible to sort search results from command line.
The sort option works exactly like the REST API sort query
parameter. It is possible to sort in ascending and descending
order. The ascending order can be sorted like:

  $ snippy search docker --sort=-created

The descending order can be sorted like:

  $ snippy search docker --sort=created
  $ snippy search docker --sort created

The equal sign is mandatory for ascending sort because the
minus sign '-' is normally reserved for command line option.

It is possible to combine multiple fields for the sort like:

  $ snippy search docker --sort=-created,-brief

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>