Fix unique operation test for valid field
The test that verifies if a requested unique field really
exists as a database column was failing. The Sqlite database
columns were not read correctly when the tool started with
an empty database. With PostgreSQL this was not working at
all.

This was not relevant before because the unique field list
is a new feature being implemented.

The way how the column names are read from database should
be more generic now.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>