Add API endpoints to query distinct attribute values
There are now /tags and /groups API endpoints. These API
endpoints allow querying unique values of tags and groups
and related counts of the unique values.

The implementation is almost generic to be used with for
example with 'links' attribute. The only problem is the
different delimiter between links compared to tags and
groups. This can be solved rather easily though.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>