Fix Travis CI Python 3.4 build and docker install
The Python 3.4 in Travis CI with pip 18.1 failed. It tried
to install docker==4.0.1 even though the python_version
in the setup.py defines the version to be 3.7.2.

This seems to be related to pip even though the root cause
is not know. The pip change notes do not seem to list this
kind of fix in pip>18.1. Also the success and failure to
install correct version from docker based on setup.py, seems
to change for some reason. That is, the pip 19.0 installed
the docker correct but then it did not with sequential try.

Because the case is not totally clear, the pip version is
not set in the pyproject.toml. Also this is only about the
Python 3.4 which is being deprecated.

The solution is to install the setup related tools in the
Travis CI before installation and tests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>