Fix Travis CI Python 3.4 build and docker install
Previous fix failed. The problem was not in setuptools or
in the pip. The problem was that the second 'docker' package
was without python_version marker. This caused the install
always to try to install the module in Python 3.4. Now all
'docker' package versions are tied to 'python_version' test.

There seem to be no way to support multiple conditions with
the python_version marker. Since the 'docker' package supports
Python 2.7 and Python >3.4, there has to be the same package
configured three times with different version variant.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>