Refactor content type classes for REST API
The Snippet, Solution and Reference specific classes were
removed because they were just empty classes inheriting
from the base class. The base class is now directly used
because all the content types can use the same.

The implementation is now more generic and does not require
creating new files when content types are added. This is
also in line with previous refactoring of the content
type classes.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>