Fix Makefile debugging options
The Makefile is now silent by default but it allows setting a
verbose mode. This can be done from command line by defining
Makefile parameter V and QUIET. These parameters as well with
all the parameters in Makefile can now be set from environment
variables.

The Makefile variable precedence is:

  1. Makefile variable defined from command line
  2. Makefile variable defined from environment variable.
  3. Makefile variable default.

There is also a help target in Makefile to give a list of most
important targets, debugging and variable examples.

These changes try to avoid forcing user to change the Makefile
itself for debugging. This also cleans a bit the Travis CI
configuration file.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>