Add SPDX license identifier for machines
The SPDX license identifier [1] was added to allow machine
reading of the license.

With the same effort, the python shebang was removed from
all except the main file. The shebang is not needed in any
other file than the main which can be run with the Python.

Fixed the tool brief description to match the same brief
that is used in GitLab project brief and in OAS API specs.

[1] https://spdx.org/licenses/

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>