Fix REST API and Falcon framework usage
The Falcon framework was used incorrectly. The API query
parameters like 'sall' or 'sgrp' were defined in the HTTP
method like 'on_get' as a method parameters. The Falcon
framework adds the API path parameters like {id} like
this but not the query parameters.

Also the query parameters assumed to be received from the
method variables were used incorrectly and uselessly. The
query parameters are stored in the 'request.params'.

This was not visible as external fault. The code was just
incorrect and misleading.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>