Fix local SQLite database file ignore for git
There was previously a trick with command [1] to ignore
the snippy.db that was locally stored. This tried to
avoid storing a binary database into Git. It is simpler
just to ignore the file by Git and Docker build.

This was early days implementation which was not good.
The code handles now situation where the file not created
when snippy is started.

Problem with the [1] was also that it was not remembered
done when Git repo was re-cloned. This also allows removing
the git hooks and the directory. The hooks did not even
work as expected.

[1] git update-index --assume-unchanged snippy/data/storage/snippy.db

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>