Add coveralls.io configuration
The coveralls.io was configured with codecov.

The Makefile target to install coverage tools was renamed
to 'install-coverage' to install multiple Python modules
and to indicate that the target is meant for generic coverage
modules.

The Travis CI configuration YAML file was also cleaned a bit
to remove duplication of script steps in each Linux job.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>