Fix --format option with export
The --format option must conditionally override the value
defined in the content 'filename' attribute. The conditions
are documented in code and for user.

The fault was that the --format option did not define the
file format of the exported file if the 'filename' attribute
was used. For example the 'filename' was set to 'snippet.mkdn'
and the --format option was defined as 'text', the output file
was still in Markdown format.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>