Fix lint warnings for Python 2
When the lint configuration files were changed, it seems that
the lint was not tested with Python 2. There were few warnings
that were likely false positives left from configuration file
change.

There was one patch [1] that is likely in Python 3 compatible
version of lint but not in the Python 2 compatible lint. The
method name length has been changed and this change applies
the same regexp from [1] because of Python 2.

The gzip module causes warning from 'compress' member only
in case of Python 2.

[1]
https://github.com/PyCQA/pylint/pull/2149
https://github.com/PyCQA/pylint/pull/2149/files

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>