Add content attribute 'languages' (breaking)
There is now new resource attribute languages. The purpose
of this field is to store the content language or languages.
The attribute supports comma separated list of languages.

This is a preparation for 'code' content category. It was
considered that a language is so central topic that it should
have it's own field and not stored for example in groups or
tags.

Even though this is marked as breaking change, the export
and import of existing content can be used to automatically
to update the content YAML files with the new attribute.

The REST API was also changed to include the new field.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>