Fix server tests with Python 2 implementation
There was an exception when string formatted port was used
with httplib.HTTPConnection in Python 2. This fix allows
now running all server tests in any supported Python version.

With this change, the parallel running of server tests was
enabled.

Removed the coverage from 'test-server' and 'test-docker'
targets because these tests run with real server. There
is no coverage to be generated from external process that
runs the server on host or in container.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>