Fix parsing one line content attributes from templates
If a text template had user defined input value split to
multiple lines with attributes that were meant to be one
liners, there was a failure to remove newlines.

Attributes like brief, name and source are meant to be
one liners. If user defines this kind of attributes to
more than one line, they are now parsed to one line
user input. The newlines are interpreted as spaces.

It is considered that in case of Markdown template it is
less likely to make a mistake to add multi-line input for
one line attribute. For example the metadata is defined
as one liners in the template with examples. Therefore
the parsing changes are not always done for the Markdown
side. It is considered case by case for Markdown parsing
errors if updates are made.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>