Edit Python module versions for development
Pytest and Sphinx were updated.

The Pytest contains change for log formatting [1]. This
change affected the indention of the logs captured with
the 'caplog'. The Logger() tests evaluated the output of
the logger with indentions which failed. The test should
be now more generic since the indention of the captured
log is now removed from asserted output. This would have
likely failed also previously if the test case filename
or the number of digits for the row number where the log
is printed would have changed.

[1] https://github.com/pytest-dev/pytest/pull/5311

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>