Add resource attribute descriptions for documents
There is now a description for each resource attribute.

The terms table format was refactored. The problem was that
the table column line lengths were not aligned nicely. There
seems to be no good way to do this without writing HTML. The
formatting like ``snippets`` will not allow aligning the
text nicely. The terms table is manually aligned to look
nice and avoid horizontal scaling of long lines in columns.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>