Fix import plugin framework with Python 2
Calling the hook function from the plugin is different
between Python 2 and Python 3. The Python 2 has a bit
strange behaviour when the hook function pointer is
assigned into a Class attribute. This cases the first
attribute to be required to be the class instance or
else the call will fail with unbound method error. This
is not fully understood but it is related to change
between Python 2 and Python 3 [1].

[1] https://stackoverflow.com/a/16244344

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>