Remove Bandit security scan configuration file
The Bandit security scan configuration was removed because
it contained only one disabled test for asserts. The scan
was previously made for 'tests' folder and that was likely
the reason why the disabled check for assert was configured.

Now the Bandit is run only for the snippy code and there
is no need to disable any tests.

This likely affects to Codacy since it runs the Bandit.
The Coday likely runs it for tests as well. But these are
configured into Codacy configuration file later.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>