Fix Codacy configuration file
The configuration file should now prevent Bandit scanning
the tests folders that contain the assert statements. The
bandit.yml file was created previously for this purpose [1]
but it was not remembered (nor hist read).

The new configuration should do the same. This allows more
detailed checks in Codacy because there is not local Bandit
configuration file.

[1] https://github.com/heilaaks/snippy/commit/c8bfa51f64f0161572b43a2f726133b148e12c97

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>