Fix ctr-c exception while import plugin is running
If user interrupted the import plugin from command line, there
was a long and cryptic exception log thrown to the user. This
was against the logging policies. Now interrupting the plugin
from command line gives nice 'NOK' error message for the user.

The exception is still logged when -vv or --debug options are
used. This allows debugging possible problem if the plugin is
stuck somewhere. The stack trace can help to debug where the
plugin was running when it was interrupted from keyboard.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>